refactor: Move update todo action to invokable class
