+++ src/Pages/DoneTodosPage.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Jobtrek\PhpSlimTodo\Actions;
+namespace Jobtrek\PhpSlimTodo\Pages;
 
 use Jobtrek\PhpSlimTodo\Database;
 use Jobtrek\PhpSlimTodo\TodoService;
@@ -8,7 +8,7 @@ use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Slim\Views\Twig;
 
-class DoneTodosAction
+class DoneTodosPage
 {
     public function __invoke(
         ServerRequestInterface $request,
